Add more headroom for ext4 app package containers.
Code in the volume daemon that creates the filesystem image to contain
an installed application package currently adds about 1.6% to the
size estimate that is passed down to it to account for filesystem
overhead. That figure is based upon an assumption that the filesystem
is of type FAT, but is insufficient for some big application packages
when the filesystem is of type ext4.

This change checks the filesystem type, and in the case where it's
ext4, adds about 6% to the input estimate to account for journaling,
etc.

Change-Id: I261fa7ee2bc7e8ea37c3d0bfd3be00c9917a6f9f
Signed-off-by: Tom Keel <thomas.keel@intel.com>